+++ src/enterprise-operations/service-registration.js
@@ -1,5 +1,6 @@
 import { ServiceCollection } from '../foundation/di/service-collection.js';
 import { EnterpriseOperationsDescriptor } from './enterprise-operations-descriptor.js';
+import { EnterpriseOperatingModelDescriptor } from './operating-model/enterprise-operating-model-descriptor.js';
 import { EnterpriseOperationsOverviewDescriptor } from './overview/enterprise-operations-overview-descriptor.js';
 
 export function addEnterpriseOperations(services) {
@@ -9,6 +10,7 @@ export function addEnterpriseOperations(services) {
 
   services.registerSingleton('EnterpriseOperationsDescriptor', () => new EnterpriseOperationsDescriptor());
   services.registerSingleton('EnterpriseOperationsOverviewDescriptor', () => new EnterpriseOperationsOverviewDescriptor());
+  services.registerSingleton('EnterpriseOperatingModelDescriptor', () => new EnterpriseOperatingModelDescriptor());
 
   return services;
 }